Adds capability to return packets back to the kernel
This patch provides two methods to return packets to the kernel
for handling as though OVS never touched them.

Firstly, we provide a bridge knob to the implicit learning switch that
is used on the 'normal' action to instead pass packets back to the
kernel.

Secondly, we provide a new flow action 'back_to_kernel' that explicitly
asks for packets to be sent back to the kernel.

To support this, we extend the datapath on kernels since 2.6.39 to be
able to pass packets back by providing a special output port number
('OVSP_NORMAL') which causes the hook to return RX_HANDLER_PASS.  We
alter ovs-dpctl to be able to parse and display these flows.

We add a flag to the userspace bridge so that it is in either OVS mode
or kernel mode. This flag is inspected when a OPFF_NORMAL action is
encountered and, if true, the datapath is instructed to use the port
'OVSP_NORMAL'.

We extend the OVSDB schema to store this flag and extend ovs-vsctl to
manipulate this flag:
      ovs-vsctl set-port-normal-mode br0 kernel
vs.
      ovs-vsctl set-port-normal-mode br0 ovs

Finally, we add the 'back_to_kernel' action to OpenFlow which triggers a
datapath output action using 'OVSP_NORMAL'.

Signed-off-by: Chris Luke <chris_luke@cable.comcast.com>